requirements.txt
discord.py
python-dotenv
aiohttp==3.9.0b0